refactor(reviews): 더보기/접기 기능 커스텀 훅 방식으로 변경
- 범용적으로 사용할 수 있도록 ui 컴포넌트에서 커스텀 훅 방식으로 변경
- useExpandableText 훅 생성 후 review card 컴포넌트 내에 적용
- reviewsComment 컴포넌트 삭제 및 관련 코드 제거
- 프로필 모달 표시 코드 제거.(api 응답 시 유저 정보 필드 부족함.)

Resolves: #233
